port_sdl_joystick: cleanup, working with all 3 controllers
