+++ README_controllers
@@ -1,0 +1,98 @@
+
+    Gamepads and joysticks
+
+  the 'gamepad' group inside the config.ini file can be used to define how gamepad elements are picked up by the game. three different joysticks (each with two axes), two hats and six buttons can be defined using the keywords shown below.
+  based on user preference, the camera angle axes (defined by the yaw and pitch) can share the same joystick as the menu navigation (nav_ns and nav_ew), because based on the current scene rendered the game provides enough context for the gamepad controller logic to switch between flight and menu navigation modes. also of note is that a solitary axis can control movement in one direction (see below for 2 examples).
+  hats can be used to either control longitudinal/transversal movement or menu navigation.
+  the integer value expected is either '1'+ for enabled items or '0' for disabled ones. an easy way of visualizing gamepad input is to run jstest, a binary that comes with the joystick package, and adding 1 to the item identifier.
+ 
+
+possible keywords:
+[gamepad]
+axis_yaw = 0 ; roll and yaw axis (initiates turns during flight)
+axis_yaw_inv = false ; false - no inversion needed, true - axis values need inversion
+axis_pitch = 0 ; pitch (moves camera up/down during flight)
+axis_pitch_inv = false
+axis_long = 0 ; longitudinal movement (aka forward / back)
+axis_long_inv = false
+axis_trans = 0 ; transversal movement (aka strafe left/right)
+axis_trans_inv = false
+axis_nav_ns = 0 ; menu navigation North / South
+axis_nav_ns_inv = false
+axis_nav_ew = 0 ; menu navigation East / West
+axis_nav_ew_inv = false
+hat_nav = 0 ; menu navigation North / South / East / West
+hat_nav_inv = false ; invert vertical axis
+hat_mov = 0 ; longitudinal and transversal movement
+hat_mov_inv = false ; invert vertical axis
+button_spell = 0 ; joystick button-press code used for displaying the spellbook
+button_minimap = 0 ; half-screen minimap
+button_fire_R = 0 ; fire right-hand spell, also right mouse button equivalent
+button_fire_L = 0 ; fire left-hand spell, also left mouse button equivalent
+button_fwd = 0 ; joystick button-press code used for forward movement
+button_back = 0 ; joystick button-press code used for backward movement
+axis_dead_zone = 2048 ; increase if _mouse_ is unresponsive
+
+
+
+    Sample setups
+
+XBOX controllers:
+
+[gamepad]
+axis_yaw = 4 ; roll and yaw - horizontal axis on the right joystick
+axis_pitch = 5 ; pitch - vertical axis on the right joystick
+axis_long = 2 ; longitudinal movement - vertical axis on the left joystick
+axis_long_inv = true ; needs inversion
+axis_trans = 1 ; transversal movement - horizontal axis on the left joystick
+hat_nav = 1 ; menu navigation on the hat
+hat_nav_inv = true ; needs inversion on the vertical axis
+button_minimap = 1 ; half-screen minimap
+button_spell = 3 ; spellbook - (X) button
+button_fire_R = 5 ; fire right-hand spell - right button
+button_fire_L = 6 ; fire left-hand spell - left button
+axis_dead_zone = 3000 ; joysticks are quite noisy and can get bumped accidentally so definitely at least 3000
+
+
+
+Logitech Attack 3 (the joystick shares flight and navigation modes, the 3-rd axis can be used for fwd-back movement)
+
+[gamepad]
+axis_yaw = 1 ; roll and yaw axis (initiates turns during flight)
+axis_yaw_inv = false ; false - no inversion needed, true - axis values need inversion
+axis_pitch = 2 ; pitch (moves camera up/down during flight)
+axis_pitch_inv = false
+;optional axis_long = 3 ; longitudinal movement (aka forward / back)
+axis_long_inv = true
+axis_nav_ns = 1 ; menu navigation North / South
+axis_nav_ew = 2 ; menu navigation East / West
+button_spell = 8 ; joystick button-press code used for displaying the spellbook
+button_fire_R = 1 ; fire right-hand spell, also right mouse button equivalent
+button_fire_L = 2 ; fire left-hand spell, also left mouse button equivalent
+button_fwd = 6 ; joystick button-press code used for forward movement
+button_back = 7 ; joystick button-press code used for backward movement
+button_minimap = 9 ; half-screen minimap
+axis_dead_zone = 2048 ; increase if _mouse_ is unresponsive
+
+
+
+Logitech Extreme 3D PRO (the joystick shares flight and navigation modes, the 4-th axis can be used for fwd-back movement)
+
+[gamepad]
+axis_yaw = 1 ; roll and yaw axis (initiates turns during flight)
+axis_yaw_inv = false ; false - no inversion needed, true - axis values need inversion
+axis_pitch = 2 ; pitch (moves camera up/down during flight)
+axis_pitch_inv = false
+; optional axis_long = 4 ; longitudinal movement (aka forward / back)
+axis_long_inv = true
+axis_nav_ns = 1 ; menu navigation North / South
+axis_nav_ew = 2 ; menu navigation East / West
+button_spell = 7 ; joystick button-press code used for displaying the spellbook
+button_minimap = 9 ; half-screen minimap
+button_fire_R = 1 ; fire right-hand spell, also right mouse button equivalent
+button_fire_L = 2 ; fire left-hand spell, also left mouse button equivalent
+button_fwd = 5 ; joystick button-press code used for forward movement
+button_back = 3 ; joystick button-press code used for backward movement
+axis_dead_zone = 2048 ; increase if _mouse_ is unresponsive
+
+
