+++ remc2/engine/read_config.h
@@ -72,6 +72,7 @@ struct gamepad_config {
 	uint8_t hat_mov_conf;
 	uint8_t hat_nav_conf;
 	uint16_t button_spell;
+	uint16_t button_minimap;
 	uint16_t button_fire_R;
 	uint16_t button_fire_L;
 	uint16_t button_fwd;
